add single pathfile flag to generate_dataset
Long time ago `generate_dataset` automatically split the dataset
into train and test set. Then I decided to that this task should
be done outside of `generate_dataset` (for example with a sweet python
script). The how stuff of generating a two separate training and tests
remained in the program. This was bad duplicated code. When you generate
create separate train and test pathfiles, it should be an easy task to
consecutive call `generate_dataset`. This commit removes all the
unnecessary duplicated code. This also clears the path for more
DataWriters especially HDF5.