refactor TrainsetGenerator
We now sample the training samples uniform over the image and then a KD-Tree is used to
find the nearest tag. If the distance is short, the `tagness` score is
high. We accept a sample based on the tagness score. That means a
the true tags are over represented. Further work need to be
done to create a command line flag for the acceptance rate.